fix(queue): start the track when play() beat its load
`Queue::play()` dropped the request when the current track's resource had
not landed yet: `PlayerImpl::play` found an empty item slot, planted
nothing in the processor, and nothing re-triggered the load when the
resource arrived. The track then sat `Loaded` and silent forever —
`playing=false`, decoded frontier 0.0, position pinned at 0.

That window is the normal case, not an edge: only the very first `play()`
in a process is slowed by starting the output stream (130-270 ms against
CoreAudio), and that delay is the only reason the load ever won the race.
Every later player in the same process reached `play()` in microseconds
and lost it — which is what made the second and later `KitharaPlayer` in
one process look silent while its graph, stream and RT node were provably
healthy.

`play()` now records the intent as a pending select on the loading track,
the same mechanism `select()` already uses for a `Pending`/`Loading`/`Slow`
track, so `spawn_apply_after_load` applies it when the resource lands.
`PlayerImpl::play` no longer announces the current item when it planted
nothing: announcing marked the index current, and the select that plants
the arriving resource would then take `select_item_with_crossfade`'s
reselecting-current path and never enqueue it.

The engine-event handling moves out of `queue/playback.rs` into
`queue/engine_events.rs`: the added arm pushed `impl Queue` past the
`single_impl_size` deny threshold, and the player-event reactions are a
separate responsibility from transport control.
